Update iptables_rules.json
If the tunnel goes down - as configured - RaspAP will try to still pass packets.  I am guessing that if someone is using a vpn, they are using it for a reason - if the tunnel breaks down for any reason, packets should be stopped completely.

Adding rule -P Forward Drop will restrict packets in event wg* or tun* goes down.